+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Bug fixes:
 * Fixed some race conditions in writing the pidfile on POSIX systems.
 * Fixed bug causing agent to stop copying logs when previously copied logs are removed
 * Disabled checking the command of the running agent process to guard against pid re-use since this was leading to some false negatives in the ``is the agent running`` checks.
+* Remove exception logging when bad response received from server
+* Print more informative message when server responds with ``server too busy``
 
 ## 2.0.13 "Moonstruck Monkey" - August 14, 2015